Extract divide into a new IntegralDomain class
Unfortunately, to still be able to provide default implementations we
have import cycles. These can be avoided by moving `IntegralDomain` into
Euclidean.hs, though it is unclear that the name would then still be
appropriate.

Cycles caused mainly by:
* Use of `WrappedIntegral`/`WrappedFractional` for `DerivingVia`
  instances for `IntegralDomain`
* `GcdDomain` now needs `IntegralDomain` context

In addition, to minimize context, added `Fractional (Complex a)`
constraint to the `Complex a` instances, which GHC dislikes.
If the required extensions are unpalatable, the constraint needs to be
replaced with a `RealFrac a` constraint.

Not all types got a `IntegralDomain` instance, since I couldn't work out
what it should be. These include the 'containers' types and `IO`.